+++ src/main.server.ts
@@ -1,0 +1,4 @@
+// Entry point for server side rendered app.
+// This file is used by angular CLI when npm run build:ssr is executed as server angular CLI config is used.
+
+export { AppServerModule } from './app/app.server.module';